Refactor cost estimation functionality
The previous implementation failed in the following case: if the user adds some
files to the context, for instance by selecting them on Dired and invoking
`gptel-context-add`, the estimate will be based on the return value of `(get
gptel-model :input-cost)`, i.e. the model selected as default. However, suppose the
user is interacting with AI in an org buffer and in this buffer a different
model is selected. After adding the files to the context, this buffer will
display the cost based on the default model rather than the model active on that
buffer.

To handle this case, we now keep track of all gptel buffers and update their
individual cost estimates every time a file is added to the context or removed
from it.