Revert "Refactor cost estimation functionality"
This reverts commit c7c4427e0615cf38c94a978914dc3565ee43ebe2.

The commit tried to fix the following problem: if the user adds some
files to the context, for instance by selecting them on Dired and invoking
`gptel-context-add`, the estimate will be based on the return value of `(get
gptel-model :input-cost)`, i.e. the model selected as default. However, suppose the
user is interacting with AI in an org buffer and in this buffer a different
model is selected. After adding the files to the context, this buffer will
display the cost based on the default model rather than the model active on that
buffer.

The problem with the fix is that it is not performant: it will update all
contexts with every file added to the context or remove from it. If one is
adding or removining a dozen or more files, the slowdown becomes annoying.

For the time being, I am leaving this problem unaddressed.